buffers no longer track state internally and act only as view to a ByteBuffer.  - all operations like position() or array() are delegated to the internal ByteBuffer  - offsets are calculated via ELEMENT_SIZE
-> buffers will never get out of sync with the internal buffer

+++ src/java/com/jogamp/common/nio/NativeBuffer.java
@@ -61,7 +61,7 @@ public interface NativeBuffer {
 
     public boolean isDirect();
 
-    public long[] array();
+    public byte[] array();
 
     public NativeBuffer rewind();
